requirements.txt
Flask==0.12.2
Flask-MySQLdb==0.2.0
Flask-WTF==0.14.2
fn==0.4.3
gunicorn==19.8.1
Jinja2==2.10
mysqlclient==1.3.12
nose==1.3.7
tornado==5.0.2
virtualenv==15.2.0
Werkzeug==0.14.1
WTForms==2.1
Flask-JSON==0.3.2
requests==2.18.4
passlib==1.7.1
pipenv==11.10.1
python-dotenv==0.8.2
Flask-Session==0.3.1
